requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
psutil
requests